+++ .gitignore
@@ -52,3 +52,4 @@ Desktop.ini
 *.swo
 *.tmp
 *.temp
+src/config.js
