+++ README.md
@@ -23,8 +23,9 @@ This project is licensed under the [MIT License](LICENSE).
    ```
 
    By default the server runs on port 8080. Adjust the command if you want a different port.
+Static files such as `index.html` and styles are now located in the `public/` folder. Application source code resides in `src/`.
 
-2. Open your browser at `http://localhost:8080/index.html` (or the port you chose) to use the application.
+2. Open your browser at `http://localhost:8080/public/index.html` (or the port you chose) to use the application.
 
 No build step is needed—serving the static files is enough to run the forum locally.
 
@@ -40,4 +41,4 @@ Create a `src/config.js` file or provide the values as environment variables whe
 - **HTTP_ENDPOINT** – URL of the GraphQL HTTP endpoint.
 - **WS_ENDPOINT** – GraphQL WebSocket endpoint for real-time updates. Typically the same host as `HTTP_ENDPOINT` using `wss://` and appending `?apiKey=YOUR_API_KEY`.
 
-When running in the browser without a bundler, copy `src/config.example.js` to `src/config.js` and fill in your actual values. Consider adding `src/config.js` to `.gitignore` if you wish to keep these secrets out of version control.
+When running in the browser without a bundler, copy `src/config.example.js` to `src/config.js` and fill in your actual values. The repository's `.gitignore` already excludes `src/config.js` so your secrets stay out of version control.
