+++ public/index.html
@@ -188,7 +188,7 @@
   <script src="https://unpkg.com/mic-recorder-to-mp3@2.2.1/dist/index.min.js"></script>
 
 
-  <script type="module" src="src/main.js"></script>
+  <script type="module" src="../src/main.js"></script>
 
 </body>
